requirements.txt
progressbar2
torch==1.6.0
torchvision==0.7
